+++ CHANGES.md
@@ -19,6 +19,7 @@ Fast and flexible PHP library for text tables.
 * dev
   * Added `Separator` row class, and `Table::addSeparator()` helper method.
   * Fixed `DuplicateColumnException` and `UnsupportedColumnAlignException` exceptions
+  * Improved documentation.
 
 
 * 2.6.0 (2023-10-12)